Allows Dispensed XP
This update allows for players to receive dispensed xp even when `block_natural_exp` is enabled. Instead of littering the ground with xp orbs that grant xp when picked up, it instead mimics the behaviour of player thrown bottles whereby the xp is granted instantly regardless of orbs. When bottle is dispensed and hits something, it finds the nearest player within a block radius of the location, the closest player receives the xp instantly. If no player is found then the xp is lost.